+++ allValidatorsWithInterface/src/test/edu/kit/scc/dem/first_steps/validators/TestPhoneNumberValidator.java
@@ -1,0 +1,100 @@
+package edu.kit.scc.dem.first_steps.validators;
+
+import edu.kit.scc.dem.first_steps.validators.impl.PhoneNumberValidator;
+import com.google.i18n.phonenumbers.NumberParseException;
+import com.google.i18n.phonenumbers.PhoneNumberUtil;
+import com.google.i18n.phonenumbers.Phonenumber;
+import org.junit.jupiter.api.Test;
+import static org.junit.Assert.*;
+import static org.junit.Assume.assumeNoException;
+
+public class TestPhoneNumberValidator {
+
+    private final PhoneNumberValidator validator = new PhoneNumberValidator();
+    private static final PhoneNumberUtil util = PhoneNumberUtil.getInstance();
+
+    @Test
+    public void validProcess(){
+        String[] args = {"+4972160828033","DE"};
+        try{
+            validator.processIsValid(args);
+        } catch (Exception e) {
+            assumeNoException(e);
+        }
+    }
+
+    @Test
+    public void onlyPossibleInSomeRegions() {
+        String[] args = {"123","DE"};
+        try{
+            validator.processIsValid(args);
+        } catch (Exception e) {
+            assumeNoException(e);
+        }
+    }
+
+    @Test
+    public void farTooShort(){
+        assertThrows(IllegalArgumentException.class, () -> {validator.isValid(new String[]{"1","DE"});});
+    }
+
+    @Test
+    public void possibleButNotValid(){
+        try{
+            Phonenumber.PhoneNumber number = util.parseAndKeepRawInput("+497216", "DE");
+            assertNotEquals(validator.isValid(new String[]{"+497216","DE"}), util.isPossibleNumber(number));
+        } catch (NumberParseException e) {
+            assumeNoException(e);
+        }
+    }
+
+    @Test
+    public void valid(){
+        assertTrue(validator.isValid(new String[]{"+497216081234","DE"}));
+    }
+
+    @Test
+    public void validWith00(){
+        assertTrue(validator.isValid(new String[]{"00497216081234","DE"}));
+    }
+
+    @Test
+    public void tooLong(){
+        assertThrows(IllegalArgumentException.class, () -> {validator.isValid(new String[]{"+49721608333333333333333","DE"});});
+    }
+
+    @Test
+    public void invalid(){
+        assertThrows(IllegalArgumentException.class, () -> {validator.isValid(new String[]{"+03245222342","DE"});});
+    }
+
+    @Test
+    public void invalidCountryCode(){
+        assertThrows(IllegalArgumentException.class, () -> {validator.isValid(new String[]{"+9991234123","DE"});});
+    }
+
+    @Test
+    public void tooShort(){
+        assertThrows(IllegalArgumentException.class, () -> {validator.isValid(new String[]{"+491","DE"});});
+    }
+
+    @Test
+    public void notANumber(){
+        assertThrows(IllegalArgumentException.class, () -> {validator.isValid(new String[]{"hello","DE"});});
+    }
+
+    @Test
+    public void noInputInProcess(){
+        assertThrows(IllegalArgumentException.class, () -> {validator.processIsValid(new String[]{});});
+    }
+
+    @Test
+    public void noInputInValidate(){
+        assertThrows(IllegalArgumentException.class, () -> {validator.isValid(new String[]{});});
+    }
+
+    @Test
+    public void invalidProcess(){
+        assertThrows(IllegalArgumentException.class, () -> {validator.processIsValid(new String[]{"", null});});
+    }
+}
